requirements.txt
# Core Dependencies
streamlit>=1.28.0
google-generativeai>=0.3.0
google-genai>=1.0.0
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# Utilities
typing-extensions>=4.8.0

# Image Processing & PDF Export
Pillow>=10.0.0
fpdf2>=2.7.0
requests>=2.31.0
